refactor(ui-grid): extract price context dialogs
